pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshcoredecoder"
version = "0.1.0"
authors = [
    {name = "Chris Davis", email = "chrisdavis2110@gmail.com"},
    {name = "Michael Hart", email = "michaelhart@michaelhart.me"}
]
description = "Complete Python implementation of the MeshCore Packet Decoder"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
keywords = ["meshcore", "decoder", "packet", "mesh", "networking"]
dependencies = [
    "pycryptodome>=3.19.0",
    "cryptography>=41.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "pylint>=2.17.0",
]

[project.urls]
Homepage = "https://github.com/chrisdavis2110/meshcore-decoder-py"
Repository = "https://github.com/chrisdavis2110/meshcore-decoder-py"
Issues = "https://github.com/chrisdavis2110/meshcore-decoder-py/issues"

[tool.setuptools]
packages = ["meshcoredecoder", "meshcoredecoder.decoder", "meshcoredecoder.crypto", "meshcoredecoder.types", "meshcoredecoder.utils", "meshcoredecoder.decoder.payload_decoders"]
py-modules = ["cli", "index"]

[project.scripts]
meshcore-decode = "cli:main"
